Implement deletion and recreation of immutable resources [EP-9145]
This implementation adds the feature of life cycle management for
immutable resources. When a request to update an immutable resource is
sent, we return an error. If the -force flag is set when running
the binary, the immutable resource is going to be deleted and recreated.

The dry run behavior reflects this update and returns an error if an
immutable resource should be updated or writes a warning if the -force
flag is enabled

Signed-off-by: Jonny Langefeld <jonny.langefeld@getcruise.com>